Check correct qid is returned for questionnaire type.
